% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_purified_model.R
\name{create_purified_model}
\alias{create_purified_model}
\title{Calculate Purified Models}
\usage{
create_purified_model(
  data,
  group_levels,
  hf_2_parameter_vector,
  flagged_items,
  param
)
}
\arguments{
\item{data}{A dataframe containing item response data and country information.}

\item{group_levels}{A vector specifying the levels for the grouping (factor), typically representing different countries.}

\item{hf_2_parameter_vector}{A string indicating the type of parameter vector to use, either "parameters_multigroup" or "parameters_promis".}

\item{flagged_items}{A vector of item names that have been flagged for exclusion in the model estimation.}

\item{param}{Additional parameters to be used if hf_2_parameter_vector is "parameters_promis".}
}
\value{
Returns the factor scores from the estimated model after the flagged items have been removed.
}
\description{
This function estimates purified models by removing flagged items and re-estimating the parameters. It is used in the context of item response theory (IRT) to address differential item functioning (DIF).
}
