% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_models.R
\name{DIF_models}
\alias{DIF_models}
\title{Estimate Differential Item Functioning (DIF) Models Including Age}
\usage{
DIF_models(item, data)
}
\arguments{
\item{item}{The name of the item variable as a string, e.g. PFM1}

\item{data}{The dataframe containing the item responses along with 'theta', 'country', and 'age' variables.}
}
\value{
A list of model objects for full DIF, uniform DIF, and non-uniform DIF including the 'age' variable.
}
\description{
This function extends the \code{DIF_models} function by including 'age' in the model estimation.
It estimates logistic regression models for differential item functioning with respect to 'theta',
'country', and 'age' for a given item.
}
\seealso{
rms::lrm() for the logistic regression model estimation.
}
