% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate Parameters}
\usage{
estimate_parameters(dat, grouping_vector)
}
\arguments{
\item{dat}{A dataframe containing item responses and country information.
It should include columns for each item and a column named 'country'.}

\item{grouping_vector}{A vector of country names to be included in the analysis.}
}
\value{
A dataframe with estimated multigroup parameters for the specified countries.
}
\description{
This function processes item response data to estimate multigroup parameters using mirt.
It filters the data for specified countries and returns a dataframe containing the estimated parameters.
}
\seealso{
mirt::multipleGroup for underlying model estimation.
}
