% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_multiverseDIF.R
\name{execute_multiverseDIF}
\alias{execute_multiverseDIF}
\title{Execute Multiverse Differential Item Functioning (DIF) Analysis}
\usage{
execute_multiverseDIF(
  criterion = c("lr", "lr_bon", "lr_ben", "beta", "CoxSnell", "Nagelkerke", "McFadden"),
  correction
)
}
\arguments{
\item{criterion}{A character vector indicating the statistical criteria to be used for DIF detection.
Options include "lr" (Likelihood Ratio), "lr_bon" (Likelihood Ratio with Bonferroni correction),
"lr_ben" (Likelihood Ratio with Benjamini-Hochberg correction), "beta" (Beta change),
"CoxSnell" (Cox-Snell pseudo R2), "Nagelkerke" (Nagelkerke pseudo R2), and "McFadden" (McFadden pseudo R2).}

\item{correction}{A parameter or method used for correction in the DIF detection process. Either "age" or "no"}
}
\value{
A list containing the results of the DIF analysis for the last iteration.
This includes the model fit, flagged items, and the criteria values used for flagging items.
The specific structure and components of the returned list should be described here.
}
\description{
This function performs a multiverse analysis for detecting differential item functioning (DIF).
It iteratively fits models, checks for DIF using various criteria, and flags items.
The process continues until the set of flagged items stabilizes.
}
