% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_promis_parameters.R
\name{prepare_promis_parameters}
\alias{prepare_promis_parameters}
\title{Prepare PROMIS Parameters for Multigroup IRT Model}
\usage{
prepare_promis_parameters(pars, pf_items, data)
}
\arguments{
\item{pars}{A dataframe containing the initial parameter set.}

\item{pf_items}{A vector of item IDs to be included in the analysis.}

\item{data}{A dataframe containing the complete dataset including country and item response data.}
}
\value{
A dataframe of the prepared parameters for the multigroup IRT model.
}
\description{
This function prepares PROMIS parameters for use with a multigroup item response theory (IRT) model. It involves filtering parameters and transforming them to the mirt convention.
}
