% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reest_model.R
\name{reest_model}
\alias{reest_model}
\title{Reestimate Model with Freed Item Parameters for Flagged Items}
\usage{
reest_model(flagged_items, dat, grouping_vector, parameters)
}
\arguments{
\item{flagged_items}{A vector of item names flagged for DIF.}

\item{dat}{Dataframe containing item response data and country information.}

\item{grouping_vector}{A vector of country names to be included in the analysis.}

\item{parameters}{A dataframe of initial model parameters.}
}
\value{
A multigroup model object with freed parameters for the flagged items.
}
\description{
This function re-estimates a multigroup model using mirt, specifically freeing
parameters for items identified with differential item functioning (DIF).
It assumes prior filtering of data for specific countries.
}
\seealso{
mirt::multipleGroup for the underlying model estimation.
}
