% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_promis_parameters.R
\name{select_promis_parameters}
\alias{select_promis_parameters}
\title{Select PROMIS Parameters}
\usage{
select_promis_parameters(parameters_promis, grouping_vector)
}
\arguments{
\item{parameters_promis}{A dataframe of the prepared parameters for the multigroup IRT model.}

\item{grouping_vector}{A vector of country names to be included in the analysis.}
}
\value{
A dataframe with PROMIS item parameters for the specified countries.
}
\description{
This function processes item response data to estimate multigroup parameters using mirt.
It filters the data for specified countries and returns a dataframe containing the estimated parameters.
}
\seealso{
mirt::multipleGroup for underlying model estimation.
}
